+++ src/Uno.UI/Graphics/SKCanvasElementFactory.skia.cs
@@ -1,0 +1,12 @@
+using System;
+using Windows.Foundation;
+using Microsoft.UI.Composition;
+
+namespace Uno.UI.Graphics;
+
+internal class SKCanvasVisualFactory : SKCanvasVisualBaseFactory
+{
+	public SKCanvasVisualFactory(object owner) { }
+
+	public SKCanvasVisualBase CreateInstance(Action<object, Size> renderCallback, Compositor compositor) => new SKCanvasVisual(renderCallback, compositor);
+}
